@@ -18,46 +18,46 @@
     "@services": "dist/services"
   },
   "dependencies": {
-    "axios": "0.19.0",
-    "babel-cli": "6.26.0",
-    "babel-core": "6.26.3",
-    "body-parser": "1.19.0",
-    "chalk": "2.4.2",
-    "cookie-parser": "1.4.4",
-    "csv": "5.1.2",
-    "d3v4": "4.2.2",
-    "domhandler": "3.0.0",
-    "domutils": "2.0.0",
-    "express": "4.17.1",
-    "express-session": "1.16.2",
-    "htmlparser": "1.7.7",
-    "htmlparser2": "4.0.0",
-    "jsdom": "15.1.1",
-    "module-alias": "2.2.1",
-    "node-csv": "0.1.2",
-    "passport": "0.4.0",
-    "passport-local": "1.0.0",
-    "passport-meetup": "1.0.0",
-    "passport-strategy": "1.0.0",
-    "rxjs": "6.5.2",
-    "serve-static": "1.14.1",
-    "ts-keycode-enum": "1.0.6",
-    "tslib": "1.10.0",
-    "tslint": "5.18.0",
-    "typescript": "3.5.3"
+    "axios": "~0.19.0",
+    "babel-cli": "~6.26.0",
+    "babel-core": "~6.26.3",
+    "body-parser": "~1.19.0",
+    "chalk": "~2.4.2",
+    "cookie-parser": "~1.4.4",
+    "csv": "~5.1.3",
+    "d3v4": "~4.2.2",
+    "domhandler": "~3.0.0",
+    "domutils": "~2.0.0",
+    "express": "~4.17.1",
+    "express-session": "~1.17.0",
+    "htmlparser": "~1.7.7",
+    "htmlparser2": "~4.0.0",
+    "jsdom": "~15.1.1",
+    "module-alias": "~2.2.2",
+    "node-csv": "~0.1.2",
+    "passport": "~0.4.0",
+    "passport-local": "~1.0.0",
+    "passport-meetup": "~1.0.0",
+    "passport-strategy": "~1.0.0",
+    "rxjs": "~6.5.3",
+    "serve-static": "~1.14.1",
+    "ts-keycode-enum": "~1.0.6",
+    "tslib": "~1.10.0",
+    "tslint": "~5.20.0",
+    "typescript": "~3.6.4"
   },
   "devDependencies": {
-    "@types/body-parser": "1.17.0",
-    "@types/cookie-parser": "1.4.1",
-    "@types/express": "4.17.0",
-    "@types/express-session": "1.15.13",
-    "@types/htmlparser2": "3.10.1",
-    "@types/jsdom": "12.2.4",
-    "@types/module-alias": "2.0.0",
-    "@types/node": "12.7.1",
-    "@types/passport": "1.0.0",
-    "@types/passport-local": "1.0.33",
-    "@types/passport-strategy": "0.2.35",
-    "@types/serve-static": "1.13.2"
+    "@types/body-parser": "~1.17.1",
+    "@types/cookie-parser": "~1.4.2",
+    "@types/express": "~4.17.1",
+    "@types/express-session": "~1.15.14",
+    "@types/htmlparser2": "~3.10.1",
+    "@types/jsdom": "~12.2.4",
+    "@types/module-alias": "~2.0.0",
+    "@types/node": "~12.7.12",
+    "@types/passport": "~1.0.1",
+    "@types/passport-local": "~1.0.33",
+    "@types/passport-strategy": "~0.2.35",
+    "@types/serve-static": "~1.13.3"
   }
 }
