@@ -1,13 +1,14 @@
 import "rxjs/add/operator/share";
-import { Observable } from "rxjs/Observable";
-import { Subject } from "rxjs/Subject";
+import {Observable} from "rxjs/Observable";
+import {Subject} from "rxjs/Subject";
 import ICommandOutput from "./i.command-output";
 
 export default abstract class CommandOutput<OutType> implements ICommandOutput<OutType> {
   // protected subject: Subject<OutType>;
   // protected observable: Observable<OutType>;
 
-  constructor(protected subject: Subject<OutType> = new Subject(), protected observable: Observable<OutType> = subject.share()) {}
+  constructor(protected subject: Subject<OutType> = new Subject(), protected observable: Observable<OutType> = subject.share()) {
+  }
 
   public get output(): Observable<OutType> {
     return this.observable;
