@@ -1,7 +1,7 @@
-import {readFileAsync} from "fs-helpers";
-import {createLog} from "node-logger";
 import {IManifest} from "@interfaces/i.manifest";
 import chalk from "chalk";
+import {readFileAsync} from "fs-helpers";
+import {createLog} from "node-logger";
 
 const log = createLog("Manifest Helpers");
 
